requirements.txt
Django==5.0.3
mysqlclient==2.1.1
djangorestframework==3.15.0
django-cors-headers==4.3.1
django-rest-knox==4.2.0
drf-yasg==1.21.7